Install argpppp only if argp-standalone is used.
Installation process is broken when argp-standalone is used:
we'd have to install argp-standalone too, but argp-standalone itself
is not easily installable. For the time being we don't need an install
target for e.g. Windows. In order to keep our CI builds running we
only have an install target if argp-standalone is not used.