Wrap up development preset.
The generator is not really optional: it needs to be either specified
or inherited. We were missing it, that's why cmake-gui did not allow
to select the preset. Added Ninja as generator, which is what we need
on Linux.

Moreover set /tmp as install directory, so we do not need to be root
to test installation.

Finally, CMakeLists.txt disables testing by default, as a convenience
for users just wanting to build and install. This is convenient because
then Catch2 is not needed, and there is not a special version of the
library built for unit testing, nor need tests to be built.
[skip ci]