Shorten preset names, add presets for using clang or gcc.
The presets dev-linux-ninja-clang and dev-linux-ninja-gcc are currently
intended mostly for use with qtcreator: I used to have two qtcreator
kits, one for clang and one for gcc. However, when my qtcreator
installation sees a CMakePresets.json it completely forgets about these
kits and does not allow me to select them. That is somewhat unfortunate,
because I like having configurations for clang and gcc at the same time
when working in qtcreator. I can have this again if I create two presets,
one for clang and one for gcc, and select both of them in qtcreator.