parse_integral: forbid calling this with TFloat=bool
If we do this with require clauses we need to do it on all overloads,
otherwise the compiler might simply choose one of the remaining overloads,
doing god knows what (e.g. infinite recursion). Not sure I like this,
a static_assert might be better, and it would also allow to produce
better error messages than a list of candidate templates, each marked
with 'the constraint was not satisfied'.