Decision: no default values for the time being.
Rationale: it is unlikely we need this for ourselves.
Moreover, if we're going to mimic Boost Program Options' behavior,
which is to not write to a variable if it has no default and its option
is not on the command line, then we can always add default values later
without it being a breaking change. Finally, not forcing people to
initialize their values is fully in the spirit of C++, so that's totally
fine [skip ci]